refactor: move types.ts to types folder
